Update sprite frames texture wrap
